alteração no desafio 1 (alterado de double para decimal meu deussssss), tentativa no req 6 com alteração req 1

+++ desafio1.sql
@@ -7,7 +7,7 @@ USE SpotifyClone;
 CREATE TABLE Plano(
 plano_id INT PRIMARY KEY AUTO_INCREMENT,
 plano_nome VARCHAR(15) NOT NULL,
-plano_preco DOUBLE NOT NULL
+plano_preco DECIMAL(5,2) NOT NULL
 ) engine = InnoDB;
 
 CREATE TABLE Usuario(
